requirements.txt
flask
pymongo
python-dotenv
flask-cors
certifi
dnspython
user_agentsgi